@@ -18,10 +18,15 @@
     "sync-externals": "npm run-script sync-schema && npm run-script sync-colors",
     "sync-schema": "cd externals/metaphysics && git fetch && git checkout origin/master && npm install && npm run dump-schema -- ../../data && rm -rf node_modules",
     "test": "jest",
-    "testing": "jest --watch"
+    "testing": "jest --watch",
+    "rndebugger": "REACT_DEBUGGER=\"open 'rndebugger://set-debugger-loc?host=localhost&port=8081' --args\" npm start"
   },
   "jest": {
-    "preset": "jest-react-native"
+    "preset": "jest-react-native",
+    "testPathIgnorePatterns": [
+      "\\.snap$",
+      "/node_modules/"
+    ]
   },
   "repository": {
     "type": "git",
@@ -54,21 +59,21 @@
   "devDependencies": {
     "@kadira/react-native-storybook": "^1.12.6",
     "apphub": "^0.5.1",
-    "babel-eslint": "^6.0.0",
-    "babel-jest": "^14.1.0",
+    "babel-eslint": "^7.1.1",
+    "babel-jest": "^17.0.2",
     "babel-polyfill": "^6.13.0",
     "babel-preset-react-native": "^1.9.0",
     "babel-relay-plugin": "https://github.com/alloy/relay/releases/download/v0.9.3/babel-relay-plugin-0.9.3.tgz",
     "concurrently": "^2.2.0",
-    "danger": "0.0.10",
-    "eslint": "^2.5.3",
-    "eslint-plugin-flow-vars": "^0.2.1",
-    "eslint-plugin-flowtype": "^2.3.1",
-    "eslint-plugin-react": "^4.2.1",
+    "danger": "^0.10.0",
+    "eslint": "^3.11.0",
+    "eslint-plugin-flow-vars": "^0.5.0",
+    "eslint-plugin-flowtype": "^2.28.2",
+    "eslint-plugin-react": "^6.7.1",
     "flow-bin": "0.32.0",
-    "jest": "^15.1.1",
-    "jest-cli": "^15.1.1",
-    "jest-react-native": "^15.0.0",
+    "jest": "^18.1.0",
+    "jest-cli": "^17.0.3",
+    "jest-react-native": "^17.0.3",
     "react-storybooks-relay-container": "^1.0.0",
     "react-test-renderer": "^15.3.0"
   },
