@@ -1,14 +1,32 @@
 /* @flow */
 
-import { danger, fail } from 'danger'
-var _ = require('lodash')
+import { danger, fail, warn } from 'danger'
+import { includes } from 'lodash'
 
 // CHANGELOG check
-const hasAppChanges = danger.git.modified_files.first((path) => { _.includes(path, 'lib/') }) !== null
-const changelogChanges = _.includes(danger.git.modified_files, 'CHANGELOG.md')
+const modifiedAppFiles = danger.git.modified_files.filter(path => path.includes('lib/'))
+const changelogChanges = danger.git.modified_files.includes('CHANGELOG.md')
 
-if (hasAppChanges && !changelogChanges) {
+if (modifiedAppFiles.length > 0 && !changelogChanges) {
   fail('No CHANGELOG added.')
+}
+
+if (!danger.github.pr.body.length) {
+  fail('Please add a description to your PR.')
+}
+
+// Warns if there are changes to package.json without changes to yarn.lock.
+const packageChanged = includes(danger.git.modified_files, 'package.json')
+const lockfileChanged = includes(danger.git.modified_files, 'yarn.lock')
+if (packageChanged && !lockfileChanged) {
+  const message = 'Changes were made to package.json, but not to yarn.lock'
+  const idea = 'Perhaps you need to run `yarn install`?'
+  warn(`${message} - <i>${idea}</i>`)
+}
+
+const someoneAssigned = danger.github.pr.assignee
+if (someoneAssigned === null) {
+  fail('Please assign someone to merge this PR, and optionally include people who should review.');
 }
 
 // Danger JS doesn't support warn yet.
