@@ -4,6 +4,7 @@
 import Relay from 'react-relay'
 import React from 'react'
 import { Image, View, StyleSheet, TouchableWithoutFeedback } from 'react-native'
+import { map } from 'lodash'
 
 import ImageView from '../opaque_image_view'
 import SerifText from '../text/serif'
@@ -21,7 +22,7 @@ class Artwork extends React.Component {
       <TouchableWithoutFeedback onPress={this.handleTap.bind(this)}>
         <View>
           <ImageView style={styles.image} aspectRatio={artwork.image.aspect_ratio} imageURL={artwork.image.url} />
-          { this.artist() }
+          { this.artists() }
           { this.artworkTitle() }
           { this.props.artwork.partner ? <SerifText style={styles.text}>{ this.props.artwork.partner.name }</SerifText> : null }
           { this.saleMessage() }
@@ -30,10 +31,14 @@ class Artwork extends React.Component {
     )
   }
 
-  artist() {
-    const artist = this.props.artwork.artist
-    if (artist) {
-      return <SerifText style={[styles.text, styles.artist]}>{artist.name}</SerifText>
+  artists() {
+    const artists = this.props.artwork.artists
+    if (artists && artists.length > 0) {
+      return (
+        <SerifText style={[styles.text, styles.artist]}>
+          {map(artists, 'name').join(', ')}
+        </SerifText>
+      )
     } else {
       return null
     }
@@ -76,9 +81,11 @@ Artwork.propTypes = {
       url: React.PropTypes.string,
       aspect_ratio: React.PropTypes.number,
     }),
-    artist: React.PropTypes.shape({
-      name: React.PropTypes.string,
-    }),
+    artists: React.PropTypes.arrayOf(
+      React.PropTypes.shape({
+        name: React.PropTypes.string,
+      }),
+    ),
     partner: React.PropTypes.shape({
       name: React.PropTypes.string,
     }),
@@ -113,7 +120,7 @@ export default Relay.createContainer(Artwork, {
           url(version: "large")
           aspect_ratio
         }
-        artist {
+        artists {
           name
         }
         partner {
