@@ -62,7 +62,8 @@ class Gene extends RootContainer {
     this.route = new Routes.Gene({
       geneID: props.geneID,
       medium: medium ? medium : '*',
-      price_range: price_range ? price_range : '*-*'
+      // The replace can be removed once metaphysics#486 is merged
+      price_range: price_range ? price_range.replace(/\.00/g, '') : '*-*'
     })
   }
 }
