@@ -17,7 +17,7 @@ class Articles extends React.Component {
         <ScrollView horizontal={true}
                     showsHorizontalScrollIndicator={false}
                     scrollsToTop={false}
-                    style={{ overflow: 'visible', marginBottom:40 }}>
+                    style={{ overflow: 'visible', marginBottom: 40 }}>
           { articles.map(article => <Article key={article.__id} article={article} style={styles.article} />) }
         </ScrollView>
       </View>
