@@ -4,7 +4,6 @@
 import React from 'react'
 import Relay from 'react-relay'
 import { NativeModules, StyleSheet, View, Dimensions } from 'react-native'
-const { ARTemporaryAPIModule } = NativeModules
 
 import Events from '../../native_modules/events'
 
@@ -38,14 +37,14 @@ class Header extends React.Component {
   }
 
   componentDidMount() {
-    ARTemporaryAPIModule.followStatusForArtist(this.props.artist._id, (error, following) => {
+    NativeModules.ARTemporaryAPIModule.followStatusForArtist(this.props.artist._id, (error, following) => {
       this.setState({ following: following })
     })
   }
 
   handleFollowChange = () => {
     const newFollowersCount: number = this.state.following ? (this.state.followersCount - 1) : (this.state.followersCount + 1)
-    ARTemporaryAPIModule.setFollowArtistStatus(!this.state.following, this.props.artist._id, (error, following) => {
+    NativeModules.ARTemporaryAPIModule.setFollowArtistStatus(!this.state.following, this.props.artist._id, (error, following) => {
       if (error) {
         console.error(error)
       } else {
