@@ -4,6 +4,7 @@
 import Relay from 'react-relay'
 import React from 'react'
 import { View, StyleSheet } from 'react-native'
+import type { LayoutEvent } from '../../../system/events'
 
 import Show from './show'
 
@@ -45,7 +46,7 @@ class ShowsList extends React.Component {
     return { width: imageWidth, height: imageHeight }
   }
 
-  numberOfColumns = (isPad: Boolean) => {
+  numberOfColumns = (isPad: boolean) => {
     if (isPad) {
       return (this.props.showSize === 'large') ? 2 : 3
     }
