@@ -2,7 +2,7 @@
 'use strict'
 
 import React from 'react'
-import { Animated, Dimensions, StyleSheet, TouchableHighlight, View } from 'react-native'
+import { Animated, StyleSheet, TouchableHighlight, View } from 'react-native'
 
 import Headline from '../text/headline'
 import colors from '../../../data/colors'
