@@ -4,7 +4,6 @@
 import React from 'react'
 import Relay from 'react-relay'
 import { NativeModules, StyleSheet, View, Dimensions } from 'react-native'
-const { ARTemporaryAPIModule } = NativeModules
 
 import Events from '../../native_modules/events'
 
@@ -29,7 +28,7 @@ class Header extends React.Component {
   }
 
   componentDidMount() {
-    ARTemporaryAPIModule.followStatusForGene(this.props.gene._id, (error, following) => {
+    NativeModules.ARTemporaryAPIModule.followStatusForGene(this.props.gene._id, (error, following) => {
       this.setState({ following: following })
     })
   }
@@ -38,17 +37,19 @@ class Header extends React.Component {
   render() {
     const gene = this.props.gene
     return (
-      <View style={{paddingTop: 20}}>
-        <Headline style={[styles.base, styles.headline]}>
-          {gene.name}
-        </Headline>
+      <View>
+        <View style={styles.header}>
+          <Headline style={styles.headline} numberOfLines={2}>
+            {gene.name}
+          </Headline>
+        </View>
         {this.renderFollowButton()}
       </View>
     )
   }
 
   handleFollowChange = () => {
-    ARTemporaryAPIModule.setFollowGeneStatus(!this.state.following, this.props.gene._id, (error, following) => {
+    NativeModules.ARTemporaryAPIModule.setFollowGeneStatus(!this.state.following, this.props.gene._id, (error, following) => {
       if (error) {
         console.error(error)
       } else {
@@ -84,19 +85,24 @@ class Header extends React.Component {
   }
 }
 
-
 const isPad = Dimensions.get('window').width > 700
 
 const styles = StyleSheet.create({
-  base: {
-    textAlign: 'center',
+  header: {
+    marginTop: 15,
+    height: 36,
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingLeft: 40,
+    paddingRight: 40,
   },
   headline: {
-    fontSize: isPad ? 20 : 14,
+    textAlign: 'center',
+    fontSize: isPad ? 20 : 14
   },
   followButton: {
     height: 40,
-    marginTop: 30
+    marginTop: 30,
   }
 })
 
