@@ -63,6 +63,35 @@ class ArtistCard extends React.Component {
     })
   }
 
+  renderMetadata() {
+    const artist = this.props.artist
+    const lines = []
+
+    lines.push(
+      <Text key={1} numberOfLines={1} style={styles.sansSerifText}>
+        {artist.name.toUpperCase()}
+      </Text>
+    )
+
+    if (artist.formatted_nationality_and_birthday) {
+      lines.push(
+        <Text key={2} numberOfLines={1} style={styles.serifText}>
+          {artist.formatted_nationality_and_birthday}
+        </Text>
+      )
+    }
+
+    if (artist.formatted_artworks_count) {
+      lines.push(
+        <Text key={3} numberOfLines={1} style={[styles.serifText, styles.serifWorksText]}>
+          {artist.formatted_artworks_count}
+        </Text>
+      )
+    }
+
+    return lines
+  }
+
   render() {
     const artist = this.props.artist
     const imageURL = artist.image && artist.image.url
@@ -76,9 +105,7 @@ class ArtistCard extends React.Component {
           <View style={styles.touchableContainer}>
             <ImageView style={styles.image} imageURL={imageURL} />
             <View style={styles.textContainer}>
-              <Text numberOfLines={1} style={styles.sansSerifText}>{artist.name.toUpperCase()}</Text>
-              <Text numberOfLines={1} style={styles.serifText}>{artist.formatted_nationality_and_birthday}</Text>
-              <Text numberOfLines={1} style={[styles.serifText, styles.serifWorksText]}>{artist.formatted_artworks_count}</Text>
+              {this.renderMetadata()}
             </View>
             <View style={styles.followButton}>
               <InvertedButton text={this.state.following ? 'Following' : 'Follow'}
@@ -155,6 +182,8 @@ const styles = StyleSheet.create({
     marginLeft: 12,
     width: 196,
     height: 30,
+    position: 'absolute',
+    bottom: 15
   }
 })
 
