@@ -3,7 +3,7 @@
 
 import Relay from 'react-relay'
 import React from 'react'
-import _ from "lodash"
+import _ from 'lodash'
 import { View, StyleSheet, TouchableHighlight, LayoutAnimation, Text, Image, Dimensions } from 'react-native'
 
 import Spinner from '../../spinner'
@@ -23,6 +23,7 @@ const additionalContentRails = [
   'current_fairs',
   'genes',
   'generic_gene',
+  'followed_artist',
 ]
 
 class ArtworkRail extends React.Component {
@@ -68,6 +69,7 @@ class ArtworkRail extends React.Component {
       case 'followed_artists':
         url = '/works-for-you'
         break
+      case 'followed_artist':
       case 'related_artists':
         url = context && context.artist.href
         break
@@ -254,6 +256,7 @@ export default Relay.createContainer(ArtworkRail, {
           ${fragments.geneFragment}
           ${fragments.auctionFragment}
           ${fragments.fairFragment}
+          ${fragments.followedArtistFragment}
         }
         results @include(if: $fetchContent) {
           ${Grid.getFragment('artworks')}
