@@ -32,7 +32,7 @@ class ArtworkRailHeader extends React.Component {
 
   constructor(props) {
     super(props)
-    this.state = { following: false }
+    this.state = { following: props.rail.key === 'followed_artist' }
   }
 
   render() {
@@ -63,7 +63,7 @@ class ArtworkRailHeader extends React.Component {
   actionButton() {
     if (this.hasAdditionalContent()) {
       return <Text style={styles.viewAllButton}> { 'View All'.toUpperCase() } </Text>
-    } else if (this.props.rail.key === 'related_artists') {
+    } else if (this.props.rail.key === 'related_artists' || this.props.rail.key === 'followed_artist') {
         return (
           <View style={styles.followButton}>
             <Button text={this.state.following ? 'Following' : 'Follow'}
@@ -137,6 +137,7 @@ export default Relay.createContainer(ArtworkRailHeader, {
           ${fragments.geneFragment}
           ${fragments.auctionFragment}
           ${fragments.fairFragment}
+          ${fragments.followedArtistFragment}
         }
       }
     `,
