@@ -48,9 +48,23 @@ function fairFragment() {
   )
 }
 
+function followedArtistFragment() {
+  return (
+    Relay.QL`
+      fragment followed_artist_context on HomePageModuleContextFollowedArtist {
+        artist {
+          href
+          id
+        }
+      }
+    `
+  )
+}
+
 export default {
   relatedArtistFragment,
   geneFragment,
   auctionFragment,
-  fairFragment
+  fairFragment,
+  followedArtistFragment,
 }
