@@ -3,10 +3,11 @@
 
 import Relay from 'react-relay'
 import React from 'react'
-import { StyleSheet, View, Dimensions } from 'react-native'
+import { StyleSheet, View } from 'react-native'
 
 import SerifText from '../text/serif'
 import RelatedArtist from './related_artist'
+import type { LayoutEvent } from '../../system/events'
 
 class RelatedArtists extends React.Component {
   state: {
@@ -16,31 +17,27 @@ class RelatedArtists extends React.Component {
 
   constructor(props) {
     super(props)
-    this.state = this.layoutState()
-  }
-
-  layoutState() : Object {
-    const width = Dimensions.get('window').width
-    const isPad = width > 700
-    const isPadHorizontal = width > 1000
-
-    // TODO: Document what these margins are based on.
-    let columnCount
-    let margins
-    if (isPad) {
-      if (isPadHorizontal) {
-        columnCount = 4
-        margins = 140
-      } else {
-        columnCount = 3
-        margins = 100
-      }
-    } else {
-      columnCount = 2
-      margins = 60
+    this.state = {
+        columns: 0,
+        imageSize: {
+          width: 1,
+          height: 1,
+        },
     }
 
-    const imageWidth = (width - margins) / columnCount
+
+  }
+
+  layoutState(currentLayout) : Object {
+    const width = currentLayout.width
+    const isPad = width > 600
+    const isPadHorizontal = width > 900
+
+    const columnCount = isPad ? (isPadHorizontal ? 4 : 3) : 2
+    const marginWidth = 20
+    const totalMargins = marginWidth * (columnCount - 1)
+
+    const imageWidth = (width - totalMargins) / columnCount
     const imageHeight = imageWidth / 1.5
 
     return {
@@ -52,8 +49,8 @@ class RelatedArtists extends React.Component {
     }
   }
 
-  onLayout = (e) => {
-    const newLayoutState = this.layoutState()
+  onLayout = (event: LayoutEvent) => {
+    const newLayoutState = this.layoutState(event.nativeEvent.layout)
     if (this.state.columns !== newLayoutState.columns) {
       this.setState(newLayoutState)
     }
@@ -61,7 +58,7 @@ class RelatedArtists extends React.Component {
 
   render() {
     return (
-      <View style={styles.container} onLayout={this.onLayout}>
+      <View style={styles.container} onLayout={this.onLayout.bind(this)}>
         <SerifText style={styles.heading}>Related Artists</SerifText>
         <View style={styles.artistContainer}>
           { this.renderArtists() }
