@@ -2,7 +2,7 @@
 'use strict'
 
 import React from 'react'
-import { StyleSheet, View, Dimensions } from 'react-native'
+import { StyleSheet, View } from 'react-native'
 
 import SwitchView from './switch_view'
 
