@@ -7,7 +7,7 @@ import { StyleSheet, Text } from 'react-native'
 export default class Headline extends React.Component {
   render() {
     return (
-      <Text style={[styles.default, this.props.style, styles.required]}>
+      <Text {...this.props} style={[styles.default, this.props.style, styles.required]} >
         {this.props.children.toUpperCase()}
       </Text>
     )
