@@ -1,61 +1,158 @@
+// @flow
+
 'use strict'
-// import Gene from '../gene'
+import 'react-native'
+import React from 'react'
+import renderer from 'react-test-renderer'
 
-  // // Wondering why this file is non-standard with imports?
-  // // Well, there is a bug which is fixed in 15.4.0
-  // // but that's not out yet, so as we only have one
-  // // non-trivial test, I can work around that easily
-  // //
-  // // @see: https://github.com/facebook/jest/issues/1353#issuecomment-249906262
+var refineCallbackPromise = () => Promise.resolve()
+jest.mock('../../native_modules/refine_callback.js', () => { return { triggerRefine: () => refineCallbackPromise() }})
 
-  // beforeEach(() => jest.resetModules())
+// Stub out these views for simplicities sake
+jest.mock('../../components/gene/header', () => 'Header')
+jest.mock('../../components/gene/artworks', () => 'Artworks')
 
-// it('is empty when *-*', () => {
-//   expect(priceRangeToHumanReadableString('*-*')).toEqual('')
-// })
+// Native view the Gene references
+jest.mock('../../components/opaque_image_view', () => 'ImageView')
+jest.mock('../../components/spinner', () => 'Spinner')
+jest.mock('../../components/switch_view', () => 'Spinner')
 
-// it('looks right when there is only a min value', () => {
-//   expect(priceRangeToHumanReadableString('50.00-*')).toEqual('Above $50')
-// })
+import Gene from '../gene'
 
-// it('looks right when there is only a max value', () => {
-//   expect(priceRangeToHumanReadableString('*-100.00')).toEqual('Below $100')
-// })
+describe('state', () => {
+  it('sets up the initial state in componentWillMount', () => {
+      const gene = new Gene()
+      gene.props = {
+        medium: 'glitch',
+        price_range: '*-*'
+      }
 
-// it('looks right when there is a max and mix value', () => {
-//   expect(priceRangeToHumanReadableString('100.00-10000.00')).toEqual('$100 - $10,000')
-// })
+      gene.componentWillMount()
 
-it('is a redundant test', () => {
-  expect(1 + 1).toEqual(2)
+      expect(gene.state).toEqual({
+        selectedTabIndex: 0,
+        showingStickyHeader: true,
+        sort: '-partner_updated_at',
+        selectedMedium: 'glitch',
+        selectedPriceRange: '*-*',
+      })
+  })
+
+  it('updates from the switch change the selectedTabIndex', ()=> {
+      const gene = new Gene()
+      const switchEvent = {
+        nativeEvent: {
+          selectedIndex: 23
+        }
+      }
+
+      gene.setState = jest.fn()
+      gene.switchSelectionDidChange(switchEvent)
+
+      expect(gene.setState).lastCalledWith({selectedTabIndex: 23})
+  })
+
+  // OK, this is a long one, but it's important.
+
+  it('updates the state with new data from Eigen', () => {
+    // Setup a Gene Component like normal
+     const gene = new Gene()
+      gene.setState = jest.fn()
+      gene.props = {
+        medium: 'glitch',
+        price_range: '*-*',
+        relay: { setVariables: jest.fn() },
+        gene: { filtered_artworks: { aggregations: []} }
+      }
+
+      // The data we expect back from Eigen when you've hit the refine button,
+      // this is a promise that Eigen would normally resolve (via the modal)
+      refineCallbackPromise = () => Promise.resolve({
+        medium: 'porcupines',
+        selectedPrice: '1000-80000',
+        sort: '-desc'
+      })
+
+      // Mount it to set up initial state
+      gene.componentWillMount()
+
+      // Then when the gene has been tapped, it returns the refine data above
+      return gene.refineTapped().then(() => {
+
+      // This should trigger new state inside the component
+      expect(gene.setState).lastCalledWith({
+        selectedMedium: 'porcupines',
+        selectedPriceRange: '1000-80000',
+        sort: '-desc'
+      })
+
+      // As well as trigger new state for Relay ( triggering a new call to metaphysics )
+      expect(gene.props.relay.setVariables).lastCalledWith({
+        medium: 'porcupines',
+        price_range: '1000-80000',
+        sort: '-desc'
+      })
+    })
+  })
 })
-  // it.skip('looks like expected', () => {
-  //   const React = require('react')  // eslint-disable-line no-unused-vars
-  //   const renderer = require('react-test-renderer')
 
-  //   const props = {
-  //       gene: {
-  //         id: 'An ID',
-  //         _id: 'a UUID',
-  //         name: 'Example Gene',
-  //         description: "Here's some text",
-  //         trending_artists: [
-  //           {
-  //             id: 'an artist',
-  //             name: 'Artist name',
-  //             counts: {
-  //               for_sale_artworks: 1,
-  //               artworks: 2
-  //             },
-  //             image: {
-  //               large_version: ''
-  //             }
-  //           }
-  //         ]
-  //       }
-  //     }
-  //   const tree = renderer.create(
-  //     <Gene geneID={props.gene.name} gene={props.gene}/>
-  //   ).toJSON()
-  //   expect(tree).toMatchSnapshot()
-  // })
+describe('handling price ranges', () => {
+  it('is empty when *-*', () => {
+    const gene = new Gene()
+    expect(gene.priceRangeToHumanReadableString('*-*')).toEqual('')
+  })
+
+  it('looks right when there is only a min value', () => {
+    const gene = new Gene()
+    expect(gene.priceRangeToHumanReadableString('50.00-*')).toEqual('Above $50')
+  })
+
+  it('looks right when there is only a max value', () => {
+    const gene = new Gene()
+    expect(gene.priceRangeToHumanReadableString('*-100.00')).toEqual('Below $100')
+  })
+
+  it('looks right when there is a max and mix value', () => {
+    const gene = new Gene()
+    expect(gene.priceRangeToHumanReadableString('100.00-10000.00')).toEqual('$100 - $10,000')
+  })
+})
+
+it('looks like expected', () => {
+  const props = {
+      gene: {
+        id: 'An ID',
+        _id: 'a UUID',
+        name: 'Example Gene',
+        description: "Here's some text",
+        filtered_artworks: {
+            total: 12,
+            aggregations: [{
+              slice: '1212',
+              counts: {
+                id: 'OK',
+                name: 'Sure',
+                count: 'Yep'
+              }
+            }]
+        },
+        trending_artists: [
+          {
+            id: 'an artist',
+            name: 'Artist name',
+            counts: {
+              for_sale_artworks: 1,
+              artworks: 2
+            },
+            image: {
+              large_version: ''
+            }
+          }
+        ]
+      }
+    }
+  const tree = renderer.create(
+    <Gene geneID={props.gene.name} medium="painting" price_range="*-100000" gene={props.gene}/>
+  ).toJSON()
+  expect(tree).toMatchSnapshot()
+})
