@@ -8,9 +8,9 @@ import _ from 'lodash'
 
 import ParallaxScrollView from 'react-native-parallax-scroll-view'
 
+import WhiteButton from '../components/buttons/flat_white'
 import Separator from '../components/separator'
 import SerifText from '../components/text/serif'
-import WhiteButton from '../components/buttons/flat_white'
 
 import About from '../components/gene/about'
 import Header from '../components/gene/header'
@@ -72,7 +72,7 @@ class Gene extends React.Component {
       selectedTabIndex: 0,
       showingStickyHeader: true,
 
-      sort: '-year',
+      sort: '-partner_updated_at',
       selectedMedium: this.props.medium,
       selectedPriceRange: this.props.price_range,
     }
@@ -121,7 +121,6 @@ class Gene extends React.Component {
     return this.selectedTabTitle() === TABS.WORKS
   }
 
-
   foregroundHeight: number = 200
 
   /** Top of the Component */
@@ -153,7 +152,7 @@ class Gene extends React.Component {
 
   refineTapped = (button) => {
     const initialSettings = {
-      sort: '-year',
+      sort: '-partner_updated_at',
       selectedMedium: this.props.medium,
       selectedPrice: this.props.price_range,
       aggregations: this.props.gene.filtered_artworks.aggregations
@@ -166,7 +165,9 @@ class Gene extends React.Component {
       aggregations: this.props.gene.filtered_artworks.aggregations
     }
 
-    Refine.triggerRefine(this, initialSettings, currentSettings).then( (newSettings) => {
+    // We're returning the promise so that it's easier
+    // to write tests with the resolved state
+    return Refine.triggerRefine(this, initialSettings, currentSettings).then( (newSettings) => {
       this.setState({
         selectedMedium: newSettings.medium,
         selectedPriceRange: newSettings.selectedPrice,
@@ -201,11 +202,14 @@ class Gene extends React.Component {
     const topMargin = this.state.showingStickyHeader ? 0 : HeaderHeight
     const separatorColor = this.state.showingStickyHeader ? 'white' : colors['gray-regular']
 
+    const refineButtonWidth = 80
+    const maxLabelWidth = Dimensions.get('window').width - (this.commonPadding * 2) - refineButtonWidth - 10
+
     return (<View style={{ backgroundColor: 'white'}}>
         <Separator style={{marginTop:topMargin, backgroundColor:separatorColor}} />
         <View style={{flexDirection: 'row', justifyContent: 'space-between', height: 26, marginTop:12, marginBottom:12, paddingLeft: this.commonPadding, paddingRight: this.commonPadding }} >
-          <SerifText style={{ fontStyle: 'italic', marginTop:4 }}>{ this.artworkQuerySummaryString() }</SerifText>
-          {/* <WhiteButton text="REFINE" style={{ height: 26, width: 80, }} onPress={this.refineTapped}/> */}
+          <SerifText style={{ fontStyle: 'italic', marginTop: 2, maxWidth: maxLabelWidth }}>{ this.artworkQuerySummaryString() }</SerifText>
+          <WhiteButton text="REFINE" style={{ height: 26, width: refineButtonWidth, }} onPress={this.refineTapped}/>
         </View>
         <Separator style={{ backgroundColor:separatorColor }}/>
       </View>)
@@ -249,8 +253,9 @@ class Gene extends React.Component {
     let items: string[] = []
     const works = this.props.gene.filtered_artworks.total.toLocaleString()
     items.push(`${works} works`)
+
     if (this.state.selectedMedium !== '*') { items.push( _.startCase(this.state.selectedMedium) ) }
-    // if (this.state.selectedPriceRange !== '*-*') { items.push( this.priceRangeToHumanReadableString(this.props.selectedPriceRange) ) }
+    if (this.state.selectedPriceRange !== '*-*') { items.push( this.priceRangeToHumanReadableString(this.state.selectedPriceRange) ) }
     return items.join(' ・ ')
   }
 
@@ -324,7 +329,7 @@ export default Relay.createContainer(Gene, {
   initialVariables: {
     medium: '*',
     price_range: '*-*',
-    sort: '-year'
+    sort: '-partner_updated_at'
   },
   fragments: {
     gene: () => Relay.QL`
