@@ -3,6 +3,8 @@
 
 import Relay from 'react-relay'
 import React from 'react'
+
+// $FlowFixMe: Flow isn't seeing the ListViewDataSource
 import { ListView, ListViewDataSource, ScrollView, RefreshControl } from 'react-native'
 
 import HeroUnits from '../components/home/hero_units'
@@ -30,7 +32,8 @@ class Home extends React.Component {
       { type: 'hero_units', data: props.home.hero_units },
     ]
     const artwork_modules = props.home.artwork_modules
-    const artist_modules = props.home.artist_modules.concat() // create a copy so we can mutate it (with `shift`)
+    const artist_modules = props.home.artist_modules && props.home.artist_modules.concat() // create a copy so we can mutate it (with `shift`)
+
     for (let i = 0; i < artwork_modules.length; i++) {
       const artwork_module = artwork_modules[i]
       rows.push({ type: 'artwork', data: artwork_module })
@@ -81,7 +84,7 @@ class Home extends React.Component {
                 }}
                 renderRow={({ type, data }, _, row) => {
                   // Offset row because we don’t store a reference to the search bar and hero units rows.
-                  const registerModule = (module) => this.state.modules[row - 2] = module
+                  const registerModule = (module) => this.state.modules[row - 2] = module  // eslint-disable-line no-return-assign
                   switch (type) {
                     case 'search_bar':
                       return <SearchBar />
